# docs/api/v1/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5078/"
    }
  ],
  "paths": {
...
    "/api/proposals/templates": {
      "post": {
        "tags": [
          "Proposals"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProposalTemplateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProposalTemplateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProposalTemplateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Proposals"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/proposals": {
      "post": {
        "tags": [
          "Proposals"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProposalRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProposalRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Proposals"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/proposals/{proposalId}": {
      "get": {
        "tags": [
          "Proposals"
        ],
        "parameters": [
          {
            "name": "proposalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/proposals/{proposalId}/pdf": {
      "get": {
        "tags": [
          "Proposals"
        ],
        "parameters": [
          {
            "name": "proposalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/proposals/{proposalId}/sign": {
      "post": {
        "tags": [
          "Proposals"
        ],
        "parameters": [
          {
            "name": "proposalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalSignRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalSignRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalSignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/proposals/{proposalId}/expire": {
      "post": {
        "tags": [
          "Proposals"
        ],
        "parameters": [
          {
            "name": "proposalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/proposals/{proposalId}/renew": {
      "post": {
        "tags": [
          "Proposals"
        ],
        "parameters": [
          {
            "name": "proposalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalRenewRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalRenewRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalRenewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/proposals/{proposalId}/reminders/force-due": {
      "post": {
        "tags": [
          "Proposals"
        ],
        "parameters": [
          {
            "name": "proposalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/proposals/{proposalId}/reminders/requeue": {
      "post": {
        "tags": [
          "Proposals"
        ],
        "parameters": [
          {
            "name": "proposalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/proposals/track/{trackingToken}": {
      "get": {
        "tags": [
          "Proposals"
        ],
        "parameters": [
          {
            "name": "trackingToken",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateProposalRequest": {
        "required": [
          "leadId",
          "title",
          "currency"
        ],
        "type": "object",
        "properties": {
          "leadId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "maxLength": 180,
            "type": "string"
          },
          "amount": {
            "maximum": 999999999,
            "minimum": 1,
            "type": "number",
            "format": "double"
          },
          "currency": {
            "maxLength": 8,
            "type": "string"
          },
          "recipientName": {
            "maxLength": 160,
            "type": "string",
            "nullable": true
          }
        }
      },
      "CreateProposalTemplateRequest": {
        "required": [
          "name",
          "displayName",
          "htmlBody"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "type": "string"
          },
          "displayName": {
            "maxLength": 160,
            "type": "string"
          },
          "htmlBody": {
            "type": "string"
          },
          "makeCurrent": {
            "type": "boolean"
          }
        }
      },
...
      "ProposalRenewRequest": {
        "type": "object",
        "properties": {
          "newExpiryDays": {
            "maximum": 90,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ProposalSignRequest": {
        "required": [
          "signerName",
          "signerEmail"
        ],
        "type": "object",
        "properties": {
          "signerName": {
            "maxLength": 160,
            "type": "string"
          },
          "signerEmail": {
            "maxLength": 320,
            "type": "string"
          }
        }
      },
...
    }
  },
...
}